rag_service/app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0

# Embedding model library
FlagEmbedding==1.2.9

# Vector Database
lancedb==0.25.3

# LLM for roadmap generation
google-generativeai==0.3.2

# Environment variable management
python-dotenv==1.0.0

# Needed for file handling + requests
python-multipart==0.0.9
requests==2.31.0